+++ test/vector.js
@@ -169,4 +169,54 @@ describe('Vector', function() {
 
     assert.strictEqual(vector.get(2), 24);
   });
+
+  it('should be possible to create a vector from an arbitrary iterable.', function() {
+    var vector = Vector.from([1, 2, 3], Uint8Array);
+
+    assert.strictEqual(vector.length, 3);
+    assert.strictEqual(vector.capacity, 3);
+    assert.deepEqual(Array.from(vector), [1, 2, 3]);
+
+    vector = Vector.from([1, 2, 3], Uint8Array, 10);
+
+    assert.strictEqual(vector.length, 3);
+    assert.strictEqual(vector.capacity, 10);
+    assert.deepEqual(Array.from(vector), [1, 2, 3]);
+
+    vector = Vector.Uint8Vector.from([1, 2, 3]);
+
+    assert.strictEqual(vector.length, 3);
+    assert.strictEqual(vector.capacity, 3);
+    assert.deepEqual(Array.from(vector), [1, 2, 3]);
+  });
+
+  it('should be possible to create a values iterator.', function() {
+    var vector = Vector.from([1, 2, 3], Uint8Array);
+
+    var iterator = vector.values();
+
+    assert.strictEqual(iterator.next().value, 1);
+    assert.strictEqual(iterator.next().value, 2);
+    assert.strictEqual(iterator.next().value, 3);
+    assert.strictEqual(iterator.next().done, true);
+  });
+
+  it('should be possible to create an entries iterator.', function() {
+    var vector = Vector.from([1, 2, 3], Uint8Array);
+
+    var iterator = vector.entries();
+
+    assert.deepEqual(iterator.next().value, [0, 1]);
+    assert.deepEqual(iterator.next().value, [1, 2]);
+    assert.deepEqual(iterator.next().value, [2, 3]);
+    assert.strictEqual(iterator.next().done, true);
+  });
+
+  it('should be possible to iterate over the vector.', function() {
+    var vector = Vector.from([1, 2, 3], Uint8Array),
+        i = 0;
+
+    for (var item of vector)
+      assert.strictEqual(item, ++i);
+  });
 });